Splitting driver list in to train and validation lists
Created a reusable function that we can use at any point.  We'll do multiple runs later
with different mixes of data and different sizes of training and validation data sets

We are ensuring that there a driver only appears in training or validation data sets to
avoid any overfitting of the data.  There are lots of images of the data, but only a
small number of drivers